+++ resourses/test_cases/phrases.txt
