+++ resourses/test_cases/stopwords.txt
